Add and use settings.php template.
When a change is made to the databse password attribute it is not reflected in settings.php.  The result is that the code base can no longer comminicate with its database and problems such as those described in mdxp/drupal-cookbook#5 arise.

This issue is partly the result of using drush-install to install drupal and not using a template to override the settins.php file.

Unfortunately drush-install does not currently allow a changes to settings.php file via the command line (as far as I'm aware).  Therefore, this issue reimplements the template process for configuring the settins.php file.  Since d7 and d6 have different settins.php formats, a template is provided for each (@todo add d8 template).

This reverts part of commit 4547328ba0043376b3626aec26030c26c97e1fce.

Conflicts:
	attributes/default.rb
	metadata.rb
	recipes/default.rb